+++ src/routes/[key]/+page.server.ts
@@ -4,7 +4,7 @@ import { getLink } from '$lib/server/links';
 export async function load({ params, platform, setHeaders }) {
     if (!platform) error(500, 'Platform not found');
 
-    const result = await getLink(platform, params.key);
+    const result = await getLink(platform, params.key.toLowerCase());
     if (!result) error(404, 'Requested link does not exist');
 
     setHeaders({